requirements.txt
numpy==1.14.6
opencv-python==4.3.0.38
Pillow==8.0.1